This is the skeleton for the Key Phrases Menu. I moved the KeyPhrase Activity stuff to it's own Activity file. Now it doesn't remember the key phrases when switching between main menu and key phrases menu.
