+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,13 @@
         android:theme="@style/Theme.VRE_App"
         tools:targetApi="31">
         <activity
+            android:name=".KeyWordsActivity"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
