+++ app/src/main/res/layout/activity_main.xml
@@ -20,7 +20,6 @@
         android:id="@+id/goToKeyPhraseMenu"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:onClick="goToKeyPhraseMenu"
         android:text="Key Phrases"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
